insight-bot/requirements.txt
selenium==4.34.2
sounddevice==0.5.2
numpy==2.3.1
pyyaml==6.0.2
pyannote.audio==3.3.2
torch==2.7.1
ffmpeg-python==0.2.0
